+++ usr/src/common/net/wanboot/boot_http.h
@@ -20,7 +20,7 @@
  * CDDL HEADER END
  */
 /*
- * Copyright 2002-2003 Sun Microsystems, Inc.  All rights reserved.
+ * Copyright 2006 Sun Microsystems, Inc.  All rights reserved.
  * Use is subject to license terms.
  */
 
@@ -115,7 +115,7 @@ int  http_set_private_key_file(http_handle_t, const char *);
 int   http_srv_connect(http_handle_t);
 int   http_head_request(http_handle_t, const char *);
 int   http_get_request(http_handle_t, const char *);
-int   http_get_range_request(http_handle_t, const char *, off_t, size_t);
+int   http_get_range_request(http_handle_t, const char *, off64_t, off64_t);
 void  http_free_respinfo(http_respinfo_t *);
 int   http_process_headers(http_handle_t, http_respinfo_t **);
 int   http_process_part_headers(http_handle_t, http_respinfo_t **);
